ninth commit

+++ proj2/gitlet/Remove.java
@@ -23,7 +23,6 @@ public class Remove {
                 }
             }
         }
-
         /* 如果这个文件，没有在暂存区里，但是这个文件是「仓库里的正式文件」
            （也就是在你最近一次 commit 的快照里、被 HEAD 指向的branch,branch指向最新一次的提交记录，
             最新提交记录着）*/
@@ -45,6 +44,8 @@ public class Remove {
 
         if (LastCommit != null) {
             if (LastCommit.HashMapBlobs.containsKey(file)) {
+                /*如果被上一个从commit追踪*/
+
                 Stage newstage;
                 if (Main.Stage.length() > 0){
                     ObjectInputStream inp = new ObjectInputStream(new FileInputStream(Main.Stage));
@@ -53,13 +54,32 @@ public class Remove {
                 }else {
                     newstage = new Stage();
                 }
-                newstage.deleteFiles.add(file);
+
+                if (newstage.stages.containsKey(file)) {
+                    newstage.stages.remove(file);
+                }
+                else {
+                    newstage.deleteFiles.add(file);
+                }
+
                 ObjectOutputStream oos2 = new ObjectOutputStream(new FileOutputStream(gitlet.Main.Stage));
                 oos2.writeObject(newstage);
                 oos2.close();
-                File currentFile = new File(Main.CWD, file);
-                if(currentFile.exists()) {
-                    currentFile.delete();
+
+                /*
+                    当且仅当被上一个追踪才删掉工作目录中的文件
+                 */
+                File lastCommitBranch = new File(Main.Branches, Utils.readContentsAsString(Main.Head));
+                File thisCommitFile = new File(Main.Objects, Utils.readContentsAsString(lastCommitBranch));
+                ObjectInputStream ois = new ObjectInputStream(new FileInputStream(thisCommitFile));
+                Commit lastCommitFile1 = (Commit) ois.readObject();
+                if (lastCommitFile1.HashMapBlobs != null && lastCommitFile1.HashMapBlobs.size() > 0) {
+                    if (lastCommitFile1.HashMapBlobs.containsKey(file)) {
+                        File currentFile = new File(Main.CWD, file);
+                        if(currentFile.exists()) {
+                            currentFile.delete();
+                        }
+                    }
                 }
             }
             else {
